pyproject.toml
[project]
name = "bike-comparator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.84.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.133.1",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]
